requirements.txt
PySimpleGUI==4.38.0
Pillow==8.2.0
pandas==1.2.3
